+++ delivery-website/src/app/app.component.html
@@ -1,5 +1,5 @@
 
-<mat-sidenav-container>
+<!-- <mat-sidenav-container>
     <mat-sidenav #sidenav >
         <app-sidenav-list (sidenavClose)="sidenav.close()"></app-sidenav-list>
     </mat-sidenav>
@@ -10,4 +10,6 @@
         </main>
     </mat-sidenav-content>
     
-</mat-sidenav-container>
+</mat-sidenav-container> -->
+
+<router-outlet></router-outlet>
