+++ delivery-website/src/app/app.module.ts
@@ -5,32 +5,28 @@ import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-
-import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { Error404Component } from './error-404/error-404.component';
-import { ModalChangeStatusComponent } from './navigation/modal-change-status/modal-change-status.component';
-import { SidenavListComponent } from './navigation/sidenav-list/sidenav-list.component';
 import { MaterialModule } from './material.module';
-import { HeaderComponent } from './navigation/header/header.component';
+import { RouterModule } from '@angular/router';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    Error404Component,
-    ModalChangeStatusComponent,
-    HeaderComponent,
-    SidenavListComponent,
   ],
   imports: [
-    BrowserModule,
-    AppRoutingModule,
     NgbModule,
     FormsModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    MaterialModule
+    MaterialModule,
+    BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
+    RouterModule.forRoot([
+      {
+        path: 'history',
+        loadChildren: () => import('./pages/deliveryman/delivery.module').then(m => m.DeliveryModule)
+      },
+    ])
   ],
   exports: [
     BrowserAnimationsModule,
