+++ delivery-website/src/app/components/error-404/error-404.component.html
@@ -9,7 +9,7 @@
                 </div>
                 <div class="error-actions row justify-content-center">
                     <div class="col-2">
-                        <a [routerLink]="['/']" class="btn btn-primary btn-lg" style="background-color: #FFBA5C; border: #FFBA5C;"><span class="glyphicon glyphicon-home"></span>
+                        <a (click)="sendToHome()" class="btn btn-primary btn-lg" style="background-color: #FFBA5C; border: #FFBA5C;"><span class="glyphicon glyphicon-home"></span>
                             Take Me Home </a>
                     </div>
                 </div>
