+++ delivery-website/src/app/components/error-404/error-404.component.ts
@@ -1,0 +1,16 @@
+import { Component } from "@angular/core";
+import { Router } from "@angular/router";
+
+@Component({
+    templateUrl: './error-404.component.html'
+})
+
+export class Error404Component {
+
+    constructor(private router: Router) {}
+
+    sendToHome() {
+        this.router.navigate(['/'])
+    }
+    
+}
