+++ delivery-website/src/app/components/navigation/sidenav-list/sidenav-list.component.css
