--- delivery-website/src/app/error-404/error-404.component.ts
@@ -1,9 +1,0 @@
-import { Component } from "@angular/core";
-
-@Component({
-    templateUrl: './error-404.component.html'
-})
-
-export class Error404Component {
-    
-}
