+++ delivery-website/src/app/models/common/apiResponse.ts
@@ -1,0 +1,10 @@
+export interface APIResponse<T> {
+    response: T;
+    success: boolean;
+    notifications: Notification[];
+}
+  
+export interface Notification {
+    code: number;
+    message: string;
+}
