+++ delivery-website/src/app/pages/deliveryman/auth/login-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { LoginComponent } from './login/login.component';
+
+const routes: Routes = [
+  {
+    path: 'login', component: LoginComponent
+  }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)],
+})
+export class LoginRoutingModule { }
