+++ delivery-website/src/app/pages/deliveryman/deliveryman-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { DeliveryHistoryComponent } from './delivery-history/delivery-history.component';
+
+const routes: Routes = [
+  {
+    path: 'history', component: DeliveryHistoryComponent
+  }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)],
+})
+export class DeliverymanRoutingModule { }
